# AuthTask/swagger.yml
openapi: 3.0.3
info:
  title: AuthTask API
  version: "1.0.0"
  description: OpenAPI specification generated from the current controllers, DTOs, and service return types.

servers:
  - url: https://localhost:7111
    description: Local HTTPS
  - url: http://localhost:5043
    description: Local HTTP

tags:
  - name: Auth
  - name: Employee

paths:
  /api/Auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      description: Requires `X-API-KEY` and is rate-limited to 5 requests per minute.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: User registered successfully.
          headers:
            X-Tracking-Id:
              $ref: "#/components/headers/XTrackingId"
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/SimpleMessage"
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleMessage"
        "400":
          $ref: "#/components/responses/BadRequestString"
        "401":
          $ref: "#/components/responses/UnauthorizedApiKeyOrAuth"
        "403":
          $ref: "#/components/responses/ForbiddenApiKey"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /api/Auth/login:
    post:
      tags:
        - Auth
      summary: Authenticate and get JWT token
      description: Requires `X-API-KEY` and is rate-limited to 5 requests per minute.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/LoginRequest"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful.
          headers:
            X-Tracking-Id:
              $ref: "#/components/headers/XTrackingId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedApiKeyOrAuth"
        "403":
          $ref: "#/components/responses/ForbiddenApiKey"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /api/Employee/update:
    put:
      tags:
        - Employee
      summary: Update employee details
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeDto"
      responses:
        "200":
          description: Employee updated successfully.
          headers:
            X-Tracking-Id:
              $ref: "#/components/headers/XTrackingId"
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/SimpleMessage"
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleMessage"
        "400":
          $ref: "#/components/responses/BadRequestString"
        "401":
          $ref: "#/components/responses/UnauthorizedBearer"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /api/Employee/{id}:
    get:
      tags:
        - Employee
      summary: Get employee by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Employee identifier.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Employee retrieved successfully.
          headers:
            X-Tracking-Id:
              $ref: "#/components/headers/XTrackingId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDto"
        "400":
          $ref: "#/components/responses/BadRequestString"
        "401":
          $ref: "#/components/responses/UnauthorizedBearer"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    delete:
      tags:
        - Employee
      summary: Delete employee by ID
      description: Requires authenticated user with `admin` role.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Employee identifier.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Employee deleted successfully.
          headers:
            X-Tracking-Id:
              $ref: "#/components/headers/XTrackingId"
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/SimpleMessage"
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleMessage"
        "400":
          $ref: "#/components/responses/BadRequestString"
        "401":
          $ref: "#/components/responses/UnauthorizedBearer"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /api/Employee/list:
    post:
      tags:
        - Employee
      summary: Get paginated employee list
      description: Requires authenticated user with `admin` role.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaginationRequest"
      responses:
        "200":
          description: Employee list retrieved successfully.
          headers:
            X-Tracking-Id:
              $ref: "#/components/headers/XTrackingId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginationResponseEmployeeDto"
        "400":
          $ref: "#/components/responses/BadRequestString"
        "401":
          $ref: "#/components/responses/UnauthorizedBearer"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "429":
          $ref: "#/components/responses/TooManyRequests"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token in `Authorization` header.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: API key required for auth endpoints.

  headers:
    XTrackingId:
      description: Request tracking identifier added by middleware.
      schema:
        type: string

  responses:
    BadRequestString:
      description: Bad request.
      headers:
        X-Tracking-Id:
          $ref: "#/components/headers/XTrackingId"
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/SimpleMessage"
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/SimpleMessage"
              - $ref: "#/components/schemas/ProblemDetails"

    UnauthorizedApiKeyOrAuth:
      description: Unauthorized due to missing API key, invalid credentials, or invalid JWT.
      headers:
        X-Tracking-Id:
          $ref: "#/components/headers/XTrackingId"
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/SimpleMessage"
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/SimpleMessage"
              - $ref: "#/components/schemas/ApiKeyErrorResponse"
              - $ref: "#/components/schemas/ProblemDetails"

    UnauthorizedBearer:
      description: Unauthorized. Missing or invalid JWT bearer token.
      headers:
        X-Tracking-Id:
          $ref: "#/components/headers/XTrackingId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"

    ForbiddenApiKey:
      description: Invalid API key.
      headers:
        X-Tracking-Id:
          $ref: "#/components/headers/XTrackingId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiKeyErrorResponse"

    ForbiddenRole:
      description: Forbidden. Authenticated user does not have required role.
      headers:
        X-Tracking-Id:
          $ref: "#/components/headers/XTrackingId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"

    TooManyRequests:
      description: Too many requests. Rate limit exceeded.
      headers:
        X-Tracking-Id:
          $ref: "#/components/headers/XTrackingId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"

  schemas:
    SimpleMessage:
      type: string
      example: User created successfully

    ApiKeyErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: API key is missing
      required:
        - message

    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: true

    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
      required:
        - email
        - password
        - role

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
      required:
        - accessToken

    UpdateEmployeeDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        departmentId:
          type: string
          format: uuid
          nullable: true
        managerId:
          type: string
          format: uuid
          nullable: true
        dateOfJoining:
          type: string
          format: date-time
        dateOfLeaving:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
      required:
        - id
        - employeeCode
        - firstName
        - lastName
        - email
        - phoneNumber
        - dateOfJoining
        - isActive

    EmployeeDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        dateOfJoining:
          type: string
          format: date-time
        dateOfLeaving:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
      required:
        - id
        - employeeCode
        - firstName
        - lastName
        - email
        - phoneNumber
        - dateOfJoining
        - isActive

    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
          minimum: 1
          example: 1
        pageSize:
          type: integer
          format: int32
          minimum: 1
          default: 10
          example: 10
        search:
          type: string
          nullable: true
          example: john
        skip:
          type: integer
          format: int32
          readOnly: true
          description: Computed server-side from `pageNumber` and `pageSize`.
      required:
        - pageNumber

    PaginationResponseEmployeeDto:
      type: object
      properties:
        totalCount:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeDto"
      required:
        - totalCount
        - items